requirements.txt
Django==4.1.5
django-environ==0.9.0
django-import-export==3.0.2
pre-commit==3.0.0
psycopg2-binary-2.9.5
sentry-sdk==1.14.0
